+++ Garage_Rental.Core/Data/Visa.cs
@@ -15,7 +15,7 @@ namespace Garage_Rental.Core.Data
         public decimal VISA_ID { get; set; }
         public string VISA_NAME { get; set; }
         public decimal VISA_NUMBER { get; set; }
-        public decimal? VISA_AMOUNT { get; set; }
+        public double? VISA_AMOUNT { get; set; }
         public DateTime? END_DATE { get; set; }
         public decimal? CVC_CVV { get; set; }
